Document the nature of the orientation rotations
This was determined by experimentation with an old printed marker
I had to hand. There's some slight oddness here in that the values
suggest the marker I have is upside down, however that seems to
disagree with other sources I've checked for the same marker.

Also worth noting that I don't have a calibration for the camera
I used, so this is based off another calibration I found for the
same resolution.